Week 3/may 19/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.8.6
beautifulsoup4==4.12.2
langchain==0.0.335
langchain-google-genai==0.0.5
